Add types and config files
