+++ package.json
@@ -1,14 +1,15 @@
 {
   "name": "mongoose-paginate-v2",
-  "version": "1.4.2",
+  "version": "5.13.0",
   "description": "A custom pagination library for Mongoose with customizable labels.",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
     "pretest": "npm run build",
-    "test": "./node_modules/.bin/mocha tests/*.js -R spec --ui bdd --timeout 9999999",
-    "build": "babel src -d dist",
+    "test": "./node_modules/.bin/mocha -r ts-node/register tests/*/*.ts -R spec --ui bdd --timeout 9999999",
+    "build": "tsc -p tsconfig.build.json",
     "lint": "./node_modules/.bin/eslint \"tests\" \"src\" \".eslintrc.js\" --fix",
-    "prettier": "./node_modules/.bin/prettier --write \"{src,tests}/**/*.js\" \"*.md\"",
+    "prettier": "./node_modules/.bin/prettier --write \"{src,tests}/**/*.ts\" \"*.md\"",
     "prepublish": "npm run test"
   },
   "husky": {
@@ -17,7 +18,7 @@
     }
   },
   "lint-staged": {
-    "*.js": [
+    "*.ts": [
       "./node_modules/.bin/eslint --fix",
       "./node_modules/.bin/prettier --write"
     ],
@@ -46,7 +47,8 @@
   ],
   "author": "Aravind NC <aravind_n_c@yahoo.co.in> (https://aravindnc.com)",
   "contributors": [
-    "Fernando López Guevara (fernando.lguevara@gmail.com)"
+    "Fernando López Guevara (fernando.lguevara@gmail.com)",
+    "Daniel Seijo Sánchez (daniseijo.dev@gmail.com)"
   ],
   "license": "MIT",
   "bugs": {
@@ -54,35 +56,26 @@
   },
   "homepage": "https://github.com/aravindnc/mongoose-paginate-v2#readme",
   "dependencies": {},
+  "peerDependencies": {
+    "mongoose": "^5.13.x"
+  },
   "devDependencies": {
-    "@babel/cli": "^7.14.5",
-    "@babel/core": "^7.14.6",
-    "@babel/plugin-proposal-class-properties": "^7.14.5",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/plugin-proposal-do-expressions": "^7.14.5",
-    "@babel/plugin-proposal-export-default-from": "^7.14.5",
-    "@babel/plugin-proposal-export-namespace-from": "^7.14.5",
-    "@babel/plugin-proposal-function-bind": "^7.14.5",
-    "@babel/plugin-proposal-function-sent": "^7.14.5",
-    "@babel/plugin-proposal-json-strings": "^7.14.5",
-    "@babel/plugin-proposal-logical-assignment-operators": "^7.14.5",
-    "@babel/plugin-proposal-nullish-coalescing-operator": "^7.14.5",
-    "@babel/plugin-proposal-numeric-separator": "^7.14.5",
-    "@babel/plugin-proposal-optional-chaining": "^7.14.5",
-    "@babel/plugin-proposal-pipeline-operator": "^7.14.5",
-    "@babel/plugin-proposal-throw-expressions": "^7.14.5",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/plugin-syntax-import-meta": "^7.10.4",
-    "@babel/preset-env": "^7.14.7",
+    "@types/chai": "^4.2.21",
+    "@types/mocha": "^9.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.29.3",
+    "@typescript-eslint/parser": "^4.29.3",
     "babel-eslint": "^10.1.0",
     "chai": "^4.3.4",
-    "eslint": "^7.30.0",
-    "eslint-config-prettier": "^6.15.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
     "husky": "^4.3.8",
     "lint-staged": "^10.5.4",
-    "mocha": "^8.4.0",
+    "mocha": "^9.1.1",
+    "mongodb": "^3.6.11",
     "mongoose": "^5.13.2",
-    "prettier": "^2.3.2"
+    "prettier": "^2.3.2",
+    "ts-node": "^10.2.1",
+    "typescript": "^4.4.2"
   },
   "engines": {
     "node": ">=4.0.0"
